Testing both Database and Eloquent providers.
